backend/requirements.txt
blinker==1.7.0
click==8.1.7
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib-metadata==7.0.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
marshmallow==3.20.2
mysql-connector-python==8.3.0
nodeenv==1.8.0
packaging==23.2
pluggy==1.4.0
pyright==1.1.349
pytest==8.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
zipp==3.17.0
